requirements.txt
discord.py
pymongo
pillow